requirements.txt
mysql-connector-python>=8.0.26
pyfiglet>=0.8.post1
requests>=2.24.0
selenium>=4.8.0